+++ pkg/log/config.go
@@ -1,5 +1,7 @@
 package log
 
+import "time"
+
 // Config  log config
 type Config struct {
 	Development       bool
@@ -7,10 +9,12 @@ type Config struct {
 	DisableStacktrace bool
 	Encoding          string
 	Level             string
-	Name              string // service name
+	ServiceName       string // service name
+	Filename          string
 	Writers           string
 	LoggerDir         string
 	LogFormatText     bool
 	LogRollingPolicy  string
 	LogBackupCount    uint
+	FlushInterval     time.Duration // default is 30s, recommend is dev or test is 1s, prod is 1m
 }