+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat: support async flush log to disk
 - feat: add re-entry and automatic renewal for redis lock
+- feat: add gorm logger
 - chore: using gorm offical plguin for tracing and metrics
 
 ## v1.9.0